wrapped guest invitation page with guestInvitationLayout
